fix: generated app not starting when using npm 8
affects: @amplicode/codegen

Generated TypeScript typings (src/gql) are using
@graphql-typed-document-node/core package.
It is a transitive dependency of @apollo/client.
When using npm 7 it is installed on top level and is
therefore usable by src/gql.
When using npm 8 it is installed in node_modules
inside @apollo/client folder and therefore is inaccessible.
As a workaround, we are adding it as a dev dependency
to src/gql itself.